update: myprofile 리팩토링
